+++ transport/transport-backend-cct/src/main/java/com/google/android/datatransport/cct/LegacyFlgDestination.java
@@ -15,14 +15,13 @@
 package com.google.android.datatransport.cct;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import com.google.android.datatransport.runtime.Destination;
 import java.io.UnsupportedEncodingException;
 
 public class LegacyFlgDestination implements Destination {
   static final String DESTINATION_NAME = "lflg";
   private static final String DEFAULT_API_KEY =
-      CctBackendFactory.mergeStrings("AzSCki82AwsLzKd5O8zo", "IayckHiZRO1EFl1aGoK");
+      StringMerger.mergeStrings("AzSCki82AwsLzKd5O8zo", "IayckHiZRO1EFl1aGoK");
 
   public static final LegacyFlgDestination DEFAULT_INSTANCE =
       new LegacyFlgDestination(DEFAULT_API_KEY);
@@ -44,29 +43,27 @@ public class LegacyFlgDestination implements Destination {
     return DESTINATION_NAME;
   }
 
-  @Nullable
+  @NonNull
   @Override
   public byte[] getExtras() {
     return encodeString(apiKey);
   }
 
-  @Nullable
+  @NonNull
   static byte[] encodeString(@NonNull String s) {
     try {
       return s.getBytes("UTF-8");
     } catch (UnsupportedEncodingException e) {
-      e.printStackTrace();
+      throw new IllegalStateException("UTF-8 encoding not found.");
     }
-    return null;
   }
 
-  @Nullable
+  @NonNull
   static String decodeExtras(@NonNull byte[] a) {
     try {
       return new String(a, "UTF-8");
     } catch (UnsupportedEncodingException e) {
-      e.printStackTrace();
+      throw new IllegalStateException("UTF-8 encoding not found.");
     }
-    return null;
   }
 }
